pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gwanjong-mcp"
version = "0.4.0"
description = "AI Growth OS — campaign-driven social engagement with UTM tracking, scheduling, and A/B experiments"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Son Seong Joon", email = "sonsj97@plateer.com" },
]
keywords = ["mcp", "social", "agent", "devto", "bluesky", "twitter", "reddit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "mcp-pipeline>=0.1.0",
    "devhub-social[all]>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
devto = []  # httpx만 필요, 코어에 포함
bluesky = ["atproto>=0.0.55"]
twitter = ["tweepy>=4.14.0"]
reddit = ["asyncpraw>=7.7.0"]
autonomous = ["anthropic>=0.40.0"]
dashboard = ["aiohttp>=3.9.0"]
all = [
    "gwanjong-mcp[devto]",
    "gwanjong-mcp[bluesky]",
    "gwanjong-mcp[twitter]",
    "gwanjong-mcp[reddit]",
    "gwanjong-mcp[autonomous]",
    "gwanjong-mcp[dashboard]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.8.0",
    "ruff>=0.5.0",
    "playwright>=1.40.0",
]

[project.scripts]
gwanjong-mcp = "gwanjong_mcp.server:main"
gwanjong-daemon = "gwanjong_mcp.daemon:main"
gwanjong-dashboard = "gwanjong_mcp.dashboard:main"
gwanjong-approval = "gwanjong_mcp.approval_cli:main"

[project.urls]
Homepage = "https://github.com/SonAIengine/gwanjong-mcp"
Repository = "https://github.com/SonAIengine/gwanjong-mcp"
Issues = "https://github.com/SonAIengine/gwanjong-mcp/issues"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "N818"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: requires network (Playwright scraping, etc.)",
]
